+++ packages/connect-multichain/src/domain/multichain/index.ts
@@ -146,5 +146,6 @@ export function getTransportType(type: string): TransportType {
 
 export * from './api/constants';
 export * from './api/infura';
+export { EIP155_CAPABILITIES_SESSION_PROPERTY } from './api/types';
 export type * from './api/types';
 export type * from './types';